Update game version and modify publishing settings
The game version has been updated from 1.20 to 1.21 in the gradle properties. The publishing settings are also modified, making the dry run false and removing the specific publishing enablements (fabric, forge and neoforge). A pinged Discord role id is also added. Changes are also made in the README.md and build.gradle files for better readability and functionality.